Fix strings longer than 256 bytes.
When converting string tensors to numpy format, we were only reading the first
byte of the size from triton and ignoring the last 3 bytes. This caused either
a std::bad_alloc exception or segfault when trying to process strings with more
than 256 characters. Fix by correctly reading the size from the triton buffer,
and add a unittest that would have caught this.