Change API to mimic the triton python_backend
This changes the python API we use in this backend to match that of the API
found in the triton python_backend. To do this we are defining a pybind11
embedded module 'triton_python_backend_utils' that exposes objects that match
the existing python_backend.

This has a couple two main advantages:

 * Our existing model.py files (for tf/pt/hugectr) can continue to work unchanged, and
  we can develop for this backend and the existing python_backend simultaneously
 * Allows us to extend the api to enable small batchsize optimizations