Fix python errors on shutdown
In certain situations on shutdown we were seeing python errors like
```'Fatal Python error: PyEval_SaveThread: NULL tstate'```. This was
mainly occuring for me when a model threw an exception on startup.

This seems like the error is that we're trying to release the GIL after
the interpreter has shutdown. Fix by not releasing the GIL after shutdown,
and use the python PyEval_*Thread api's instead of the pybind scoped_gil_acquire
guard.