Fix error handling (#44)
We were seeing a problem where tritonserver would segfault if the python model
threw an exception - but since the segfault occurred after the error response
was sent the existing unittest worked. The problem seems to be returning an
error from TRITONBACKEND_ModelInstanceExecute causes tritonserver to try
to also send an error response back, but we've already closed the request
it fails. Not closing the request on error avoids the segfault, but causes
problems when shutting down tritonserver (it times out on shutdown since there
are still pending requests). So fix by returning success, which is fine since
we've already returned the error to the client.

Also change the unittest to detect this problem by running the error request multiple
times.